orchestrator/requirements.txt

autogen-agentchat>=0.2.0
openai>=1.40.0
httpx>=0.27.0
sqlmodel>=0.0.21
SQLAlchemy>=2.0
pytest>=8.3
rich>=13.7
